Added: Removing ingredients from meals; Fixed: Getting macros of the meals method

+++ Models/MealVM.cs
@@ -8,13 +8,13 @@ namespace TrainingPlanApp.Web.Models
         [Display(Name = "Name")]
         public string? Name { get; set; }
         [Display(Name = "Amount of kcal")]
-        public int? Kcal { get; set; }
+        public double? Kcal { get; set; }
         [Display(Name = "Amount of proteins")]
-        public int? Proteins { get; set; }
+        public double? Proteins { get; set; }
         [Display(Name = "Amount of fats")]
-        public int? Fats { get; set; }
+        public double? Fats { get; set; }
         [Display(Name = "Amount of carbs")]
-        public int? Carbohydrates { get; set; }
+        public double? Carbohydrates { get; set; }
         public List<int?>? IngredientIds { get; set; }
         public List<int?>? IngredientQuantities { get; set; }
     }
